editoast: move get_rolling_stock_usage sql request from views to modelsv2
