+++ editoast/src/modelsv2/rolling_stock_model/sql/get_power_restrictions.sql
@@ -1,0 +1,2 @@
+SELECT DISTINCT jsonb_object_keys(power_restrictions) AS power_restriction
+FROM rolling_stock
