+++ editoast/src/modelsv2/rolling_stock_model/sql/get_train_schedules_with_scenario.sql
